+++ src/lib/murmur.c
@@ -47,8 +47,8 @@ uint32_t murmur4(const uint8_t * data, uint8_t len)
 
     switch(len & 3)
     {
-        case 3: k1 ^= tail[2] << 16;
-        case 2: k1 ^= tail[1] << 8;
+        case 3: k1 ^= (uint32_t) tail[2] << 16;
+        case 2: k1 ^= (uint32_t) tail[1] << 8;
         case 1: k1 ^= tail[0];
               k1 *= c1; k1 = rotl32(k1,15); k1 *= c2; h1 ^= k1;
     }
