+++ src/util/util.c
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include "util.h"
 
-char *util_hex(uint8_t *in, uint32_t len, char *out)
+char *util_hex(uint8_t *in, size_t len, char *out)
 {
     uint32_t j;
     char *c = out;
@@ -38,7 +38,7 @@ char hexcode(char x)
     }
 }
 
-uint8_t *util_unhex(char *in, uint32_t len, uint8_t *out)
+uint8_t *util_unhex(char *in, size_t len, uint8_t *out)
 {
   uint32_t j;
   uint8_t *c = out;
@@ -86,9 +86,9 @@ int _util_sort_alpha(void *arg, const void *a, const void *b)
 /* This code is derived from a public domain shell sort routine by
  * Ray Gardner and found in Bob Stout's snippets collection. */
 
-void util_sort(void *base, int nel, int width, int (*comp)(void *, const void *, const void *), void *arg)
+void util_sort(void *base, unsigned int nel, unsigned int width, int (*comp)(void *, const void *, const void *), void *arg)
 {
-  int wgap, i, j, k;
+  unsigned int wgap, i, j, k;
   char tmp;
 
   // added this for default alpha sort
